rewrite user status functionality
