+++ app/controllers/admin/users_controller.rb
@@ -2,7 +2,6 @@ module Admin
   class UsersController < Admin::ApplicationController
     def update
       user = User.find(params[:id])
-      UserRoleService.new(user, current_user.id).check_for_roles(params[:user])
       if user.errors.messages.blank? && user.update(user_params)
         flash[:notice] = "User successfully updated"
         redirect_to "/admin/users/#{params[:id]}"
