--- app/fields/user_analytics_field.rb
@@ -1,7 +1,0 @@
-require "administrate/field/base"
-
-class UserAnalyticsField < Administrate::Field::Base
-  def to_s
-    data
-  end
-end
