--- app/fields/user_scholar_field.rb
@@ -1,7 +1,0 @@
-require "administrate/field/base"
-
-class UserScholarField < Administrate::Field::Base
-  def to_s
-    data
-  end
-end
