+++ app/services/moderator/manage_activity_and_roles.rb
@@ -1,0 +1,116 @@
+module Moderator
+  class ManageActivityAndRoles
+    attr_reader :user, :admin, :user_params
+
+    def initialize(args)
+      @user = args[:user]
+      @admin = args[:admin]
+      @user_params = args[:user_params]
+    end
+
+    def self.handle_user_roles(admin:, user:, user_params:)
+      new(user: user, admin: admin, user_params: user_params).update_roles
+    end
+
+    def self.handle_user_privileges(admin:, user:, user_params:)
+      new(user: user, admin: admin, user_params: user_params).update_privileges
+    end
+
+    def delete_comments
+      return unless user.comments.any?
+
+      user.comments.find_each do |comment|
+        comment.reactions.delete_all
+        CacheBuster.new.bust_comment(comment.commentable, user.username)
+        comment.delete
+      end
+    end
+
+    def delete_articles
+      return unless user.articles.any?
+
+      user.articles.find_each do |article|
+        article.reactions.delete_all
+        article.comments.find_each do |comment|
+          comment.reactions.delete_all
+          CacheBuster.new.bust_comment(comment.commentable, comment.user.username)
+          comment.delete
+        end
+        CacheBuster.new.bust_article(article)
+        article.remove_algolia_index
+        article.delete
+      end
+    end
+
+    def delete_user_activity
+      user.notifications.delete_all
+      user.reactions.delete_all
+      user.follows.delete_all
+      Follow.where(followable_id: user.id, followable_type: "User").delete_all
+      user.chat_channel_memberships.delete_all
+      user.mentions.delete_all
+      user.badge_achievements.delete_all
+      user.github_repos.delete_all
+      delete_comments
+      delete_articles
+    end
+
+    def remove_privileges
+      @user.remove_role :trusted
+      @user.remove_role :video_permission
+      @user.remove_role :tag_moderator
+    end
+
+    def add_note
+      return unless !user_params[:note].blank?
+
+      create_note("misc_note", user_params[:note])
+    end
+
+    def create_note(reason, content)
+      Note.create(
+        author_id: admin.id,
+        noteable_id: @user.id,
+        noteable_type: "User",
+        reason: reason,
+        content: content,
+      )
+    end
+
+    def handle_user_status(role, note)
+      case role
+      when "Ban"
+        user.add_role :banned
+        remove_privileges
+      when "Warn"
+        user.add_role :warned
+        user.remove_role :banned
+        remove_privileges
+      when "Comment Ban"
+        user.add_role :comment_banned
+        user.remove_role :banned
+        remove_privileges
+      when "Good Standing"
+        user.remove_role :banned if user.banned
+        user.remove_role :warned if user.warned
+        user.remove_role :comment_banned if user.comment_banned
+      when "Trusted"
+        user.add_role :trusted
+      when "Spammer"
+        user.add_role :banned
+        remove_privileges
+      end
+      create_note(role, note)
+      update_trusted_cache
+    end
+
+    def update_trusted_cache
+      Rails.cache.delete("user-#{@user.id}/has_trusted_role")
+      @user.trusted
+    end
+
+    def update_roles
+      handle_user_status(user_params[:user_status], user_params[:note_for_current_role])
+    end
+  end
+end
