--- app/views/fields/user_scholar_field/_form.html.erb
@@ -1,14 +1,0 @@
-<div class="field-unit__label">
-  <%= f.label field.attribute %>
-  <sub>Defaults to 1 year. Clear the date for no expiration.</sub>
-</div>
-<div class="field-unit__field" style="width: 2em;">
-  <%= f.check_box field.attribute %>
-</div>
-<%= f.date_field "workshop_expiration", value: 1.year.from_now.strftime("%Y-%m-%d"), style: "width: 25%;" %>
-<div style="margin-left: 1em;">
-  <%= f.label "Send Scholarship Email?" %>
-</div>
-<div style="margin-left: 1em;">
-  <%= f.check_box "scholar_email" %>
-</div>
