--- app/views/fields/user_scholar_field/_index.html.erb
@@ -1,1 +1,0 @@
-<%= field.to_s %>
