--- app/views/fields/user_scholar_field/_show.html.erb
@@ -1,1 +1,0 @@
-<%= field.to_s %>
